Make clang happy.

+++ src/mcts/search.cc
@@ -1252,7 +1252,7 @@ void SearchWorker::GatherMinibatch2() {
     if (cur_n > params_.GetBatchLimitStart() && cur_n < params_.GetBatchLimitEnd()) {
       in_flight_threshold =
           mix(params_.GetBatchLimitMax(), 0,
-              std::powf((static_cast<float>(cur_n) - params_.GetBatchLimitStart()) /
+              std::pow((static_cast<float>(cur_n) - params_.GetBatchLimitStart()) /
                   (params_.GetBatchLimitEnd() - params_.GetBatchLimitStart()), params_.GetBatchLimitPower()));
     } else if (cur_n >= params_.GetBatchLimitEnd()) {
       in_flight_threshold = params_.GetBatchLimitMax();
